Organizing and creating separate styles.js
